refactor: 型として依存すべきではない要素について, パッケージ構造を変更
モジュール外では型として依存することも, 継承することも認められていないインターフェース:
FloatingRandomGenerator, IntegerRandomGenerator
について, not exported なパッケージに移動した.
Javadocとしては見えるようにした.
