+++ public.package.txt
@@ -1,4 +1,5 @@
 matsu.num.statistics.random,
+matsu.num.statistics.random.accomp,
 matsu.num.statistics.random.lib,
 matsu.num.statistics.random.mix,
 matsu.num.statistics.random.service,
