+++ src/matsu/num/statistics/random/ArcsineRnd.java
@@ -10,6 +10,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.FloatingRandomGenerator;
+
 /**
  * <p>
  * 逆正弦 (Arcsine) 分布に従う乱数発生器を扱う.
