+++ src/matsu/num/statistics/random/CategoricalRnd.java
@@ -9,6 +9,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.IntegerRandomGenerator;
+
 /**
  * <p>
  * カテゴリカル分布に従う乱数発生器を扱う.
