+++ src/matsu/num/statistics/random/ExponentialRnd.java
@@ -9,6 +9,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.FloatingRandomGenerator;
+
 /**
  * <p>
  * 標準指数分布に従う乱数発生器を扱う.
