+++ src/matsu/num/statistics/random/NormalRnd.java
@@ -9,6 +9,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.FloatingRandomGenerator;
+
 /**
  * <p>
  * 標準正規分布に従う乱数発生器を扱う.
