+++ src/matsu/num/statistics/random/PoissonRnd.java
@@ -9,6 +9,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.IntegerRandomGenerator;
+
 /**
  * <p>
  * Poisson 分布に従う乱数発生器を扱う.
