+++ src/matsu/num/statistics/random/TDistributionRnd.java
@@ -9,6 +9,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.FloatingRandomGenerator;
+
 /**
  * <p>
  * Student-t 分布に従う乱数発生器を扱う.
