+++ src/matsu/num/statistics/random/YuleSimonRnd.java
@@ -10,6 +10,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.IntegerRandomGenerator;
+
 /**
  * <p>
  * Yule-Simon 分布に従う乱数発生器を扱う.
