+++ src/matsu/num/statistics/random/ZetaRnd.java
@@ -10,6 +10,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.IntegerRandomGenerator;
+
 /**
  * <p>
  * ゼータ分布に従う乱数発生器を扱う.
