+++ src/matsu/num/statistics/random/accomp/package-info.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright © 2025 Matsuura Y.
+ * 
+ * This software is released under the MIT License.
+ * http://opensource.org/licenses/mit-license.php
+ */
+
+/**
+ * (not exported)
+ * {@link matsu.num.statistics.random}
+ * パッケージと密結合したサブパッケージ.
+ * 
+ * <p>
+ * このパッケージはモジュールとしては非公開であり, {@code exports} されない.
+ * </p>
+ */
+package matsu.num.statistics.random.accomp;
